+++ gui/__init__.py
@@ -1,0 +1,1 @@
+"""LocallyFPS desktop interface."""